% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{createDisjointGeneLoci}
\alias{createDisjointGeneLoci}
\title{createDisjointGeneLoci}
\usage{
createDisjointGeneLoci(dg)
}
\arguments{
\item{dg}{Object from createDisjointGenes}
}
\value{
finds overlapping gene regions and merges them into a gene loci.
}
\description{
createDisjointGeneLoci
}

