% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{createDisjointMap}
\alias{createDisjointMap}
\title{createDisjointMap}
\usage{
createDisjointMap(dt, disjoint, ignore.strand = FALSE)
}
\arguments{
\item{dt}{Object of class "Transcripts"}

\item{disjoint}{Object from disjoinExons()}

\item{ignore.strand}{was disjoint created ignoring strand?}
}
\value{
a map between the transcript_id of dt to  to the disjoint exons of disjoint
}
\description{
createDisjointMap
}
\seealso{
disjoinExons
}

