% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{createDisjointTranscripts}
\alias{createDisjointTranscripts}
\title{createDisjointTranscripts}
\usage{
createDisjointTranscripts(dt, disjoint, map)
}
\arguments{
\item{dt}{Object of class "Transcripts"}

\item{disjoint}{Object from disjoinExons()}

\item{map}{Object from createDisjointMap()}
}
\value{
exons in dt turned into sub-exon disjoint counting bins organized by transcript_id
}
\description{
createDisjointTranscripts
}
\seealso{
disjointExons, createDisjointMap
}

