% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{disjoinExons}
\alias{disjoinExons}
\title{disjoinExon}
\usage{
disjoinExons(dt, ignore.strand = FALSE, id_lead = "dis")
}
\arguments{
\item{dt}{disjoint_transcripts from a gtf_data object}

\item{ignore.strand}{should disjoint counting bins be caculated by ignoring strand (as ooposed to over-lapping across strands?)}

\item{id_lead}{leading string for a disjoint exon key field.}
}
\value{
disjoint genome exon data.table
}
\description{
disjoinExon
}

