% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{featureExactTest}
\alias{featureExactTest}
\title{featureExactTest}
\usage{
featureExactTest(counts, treatments, field_disp = NULL,
  has_sim_truth = TRUE, feature_type = "gene")
}
\arguments{
\item{treatments}{vector indicating sample groups.  names(treatments) should be feature counts in gene_counts.}

\item{field_disp}{Use this field as tag.wise dispersion instead of asking edgeR for it.  NULL indicates use edgeR.}

\item{gene_counts}{data.table of feature counts}
}
\description{
featureExactTest
}

