% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{fixTypes}
\alias{fixTypes}
\title{fixTypes}
\usage{
fixTypes(d)
}
\arguments{
\item{d}{data.table on which to fix up types to a standard representatation for genomic data}
}
\description{
fixTypes
}

