% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{getPCExons}
\alias{getPCExons}
\title{getPCExons}
\usage{
getPCExons(gtf, keep_seqnames = c(1:22, "X", "Y"))
}
\arguments{
\item{gtf}{Transcripts data object}

\item{keep_seqnames}{which chromosomes to keep in the gtf}
}
\value{
gtf with only protein coding exon data for keep_seqname chromosomes.
}
\description{
getPCExons
}

