% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{simDisjointGenomeNonParametric}
\alias{simDisjointGenomeNonParametric}
\title{simDisjointGenomeNonParametric}
\usage{
simDisjointGenomeNonParametric(gtf_data, gene_dist, treatment, E_lib_size,
  de_prob = 0.5, min_transcripts = 1, max_transcripts = Inf,
  group_transcripts_id = "gene_locus_id", only_sim_trans = FALSE,
  force_de = NULL)
}
\arguments{
\item{gtf_data}{Object from gtfToSimData}

\item{gene_dist}{data.frame from which to sample with replacement b_0, b_1, and dispersion triplets.   Column names should be those.}

\item{treatment}{vector of numeric 1 and 2 indicating sample assignments.}

\item{E_lib_size}{vector indicating expected library size}

\item{de_prob}{probability that a simulated transcript will be differentially expressed between treatments}

\item{min_transcripts}{minimum number of transcripts a gene_locus_id should have to be included in the simulation}

\item{max_transcripts}{maximum number of transcripts allowed per gene_locus_id (excess transcripts are removed and they are sampled at random for simulation).}

\item{group_transcripts_id}{what field should transcripts be grouped by for min_transcripts/max_transcripts}

\item{force_de}{set of gene names to force into the differentially expressed category}

\item{only_sim_transcripts}{Do not simulate the full disjoint exon/genome?}
}
\description{
simDisjointGenomeNonParametric
}

